chore(move): move `internal/proxy/config` to `api/proxy/config`
In the Terraform provider we recently added support to set the global DefaultConfig, and it was
moved to internal in Juju 4. I'm moving it back.
