+++ api/apiclient.go
@@ -31,13 +31,13 @@ import (
 	"gopkg.in/retry.v1"
 
 	"github.com/juju/juju/api/base"
+	proxy "github.com/juju/juju/api/proxy/config"
 	"github.com/juju/juju/core/facades"
 	"github.com/juju/juju/core/network"
 	jujuversion "github.com/juju/juju/core/version"
 	internalerrors "github.com/juju/juju/internal/errors"
 	jujuhttp "github.com/juju/juju/internal/http"
 	internallogger "github.com/juju/juju/internal/logger"
-	proxy "github.com/juju/juju/internal/proxy/config"
 	"github.com/juju/juju/rpc"
 	"github.com/juju/juju/rpc/jsoncodec"
 	"github.com/juju/juju/rpc/params"
