+++ api/package_test.go
@@ -66,8 +66,8 @@ func (*ImportSuite) TestImports(c *tc.C) {
 		"internal/featureflag",
 		"internal/http",
 		"internal/logger",
+		"api/proxy/config",
 		"internal/proxy",
-		"internal/proxy/config",
 		"rpc/rpcreflect",
 		"internal/storage",
 		"internal/stringcompare",
