+++ cmd/containeragent/initialize/package_test.go
@@ -95,8 +95,8 @@ func (*importSuite) TestImports(c *tc.C) {
 		"internal/logger",
 		"internal/password",
 		"internal/pki",
+		"api/proxy/config",
 		"internal/proxy",
-		"internal/proxy/config",
 		"internal/provider/kubernetes/constants",
 		"rpc/rpcreflect",
 		"internal/service/pebble/identity",
