+++ cmd/containeragent/main_nix.go
@@ -20,6 +20,7 @@ import (
 	proxyutils "github.com/juju/proxy"
 
 	"github.com/juju/juju/agent/introspect"
+	proxy "github.com/juju/juju/api/proxy/config"
 	jujucmd "github.com/juju/juju/cmd"
 	"github.com/juju/juju/cmd/cmd"
 	"github.com/juju/juju/cmd/containeragent/config"
@@ -29,7 +30,6 @@ import (
 	"github.com/juju/juju/core/machinelock"
 	"github.com/juju/juju/internal/featureflag"
 	internallogger "github.com/juju/juju/internal/logger"
-	proxy "github.com/juju/juju/internal/proxy/config"
 	_ "github.com/juju/juju/internal/secrets/provider/all" // Import the secret providers.
 	"github.com/juju/juju/internal/worker/logsender"
 	"github.com/juju/juju/juju/names"
