+++ cmd/jujud-controller/agent/machine/manifolds.go
@@ -25,6 +25,7 @@ import (
 	"github.com/juju/juju/api/base"
 	"github.com/juju/juju/api/controller/crosscontroller"
 	"github.com/juju/juju/api/macaroon"
+	proxyconfig "github.com/juju/juju/api/proxy/config"
 	"github.com/juju/juju/caas"
 	"github.com/juju/juju/cmd/jujud-controller/util"
 	"github.com/juju/juju/core/flightrecorder"
@@ -45,7 +46,6 @@ import (
 	internallease "github.com/juju/juju/internal/lease"
 	internallogger "github.com/juju/juju/internal/logger"
 	internalobjectstore "github.com/juju/juju/internal/objectstore"
-	proxyconfig "github.com/juju/juju/internal/proxy/config"
 	"github.com/juju/juju/internal/s3client"
 	"github.com/juju/juju/internal/simplestreams"
 	sshimporter "github.com/juju/juju/internal/ssh/importer"
