+++ cmd/jujud/agent/machine/manifolds.go
@@ -21,6 +21,7 @@ import (
 	"github.com/juju/juju/agent/engine"
 	"github.com/juju/juju/api"
 	"github.com/juju/juju/api/base"
+	proxyconfig "github.com/juju/juju/api/proxy/config"
 	"github.com/juju/juju/core/flightrecorder"
 	"github.com/juju/juju/core/instance"
 	corelogger "github.com/juju/juju/core/logger"
@@ -32,7 +33,6 @@ import (
 	containerbroker "github.com/juju/juju/internal/container/broker"
 	"github.com/juju/juju/internal/container/lxd"
 	internallogger "github.com/juju/juju/internal/logger"
-	proxyconfig "github.com/juju/juju/internal/proxy/config"
 	"github.com/juju/juju/internal/upgrades"
 	jupgradesteps "github.com/juju/juju/internal/upgradesteps"
 	jworker "github.com/juju/juju/internal/worker"
